V3.11 added support for sequences, fixes for Hue lights
On the Harmony App you can now create sequences and start those (issue 11). With this V3.11 you can start those sequences with a single call_action.
For Hue lights some Vera issue caused failures on RGB lights, this is now fixed. On openLuup you need to latest development branch for this to work properly.
A fix for saving the Devices Button settings.